refactor(application): extract metadata_builder.rs and component_builder.rs [#314-2/5]
Extract build_metadata and build_components/build_component methods from
sbom_read_model_builder/mod.rs into dedicated sub-builder files.

- Create metadata_builder.rs with build_metadata as pub(super) free function
- Create component_builder.rs with build_components as pub(super) free function
  and build_component as private helper
- Update mod.rs to declare new modules and delegate to free functions
- Remove no-longer-needed imports (LicenseView, MetadataComponentView,
  SbomMetadataView, spdx_license_map) from mod.rs
- Update tests to call metadata_builder:: and component_builder:: directly

Closes #338
